requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1

# Background tasks
celery==5.3.4
redis==5.0.1

# Lunar calendar
lunardate==0.2.0

# Google Calendar API & OAuth
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4

# Session management
itsdangerous==2.1.2

# Notifications
python-telegram-bot==20.7
# SMTP email only - no mailgun needed

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
